Moved FPSCounter to components.
